% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_params.R
\name{decode_tree_em}
\alias{decode_tree_em}
\title{Decode tree using EM algorithm}
\usage{
decode_tree_em(tn, A, liks, ncores = 1)
}
\arguments{
\item{tn}{phylogenetic tree}

\item{A}{transition matrix}

\item{liks}{leaf likelihoods}

\item{ncores}{number of cores to use}
}
\value{
list of logZ, edge beliefs, and node beliefs
}
\description{
Decode tree using EM algorithm
}
