% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_root}
\alias{find_root}
\title{Find the root node ID of a rooted phylogeny.}
\usage{
find_root(tree)
}
\arguments{
\item{tree}{A rooted \code{phylo} object.}
}
\value{
Integer node ID for the root.
}
\description{
The root is identified as the node index that appears as a parent in
\code{tree$edge[, 1]} but never as a child in \code{tree$edge[, 2]}.
}
