% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_root_split}
\alias{find_root_split}
\title{Get the clade node IDs adjacent to the root.}
\usage{
find_root_split(tree)
}
\arguments{
\item{tree}{A rooted \code{phylo} object.}
}
\value{
Integer vector of root child node IDs.
}
\description{
Returns the child node IDs of the root (the clades immediately below the
root). For a rooted binary tree this is typically length 2.
}
