% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_params.R
\name{fit_params_em}
\alias{fit_params_em}
\title{Fit tree parameters using EM algorithm}
\usage{
fit_params_em(
  tree_fit,
  amat,
  dmat,
  initial_params = c(ngen = 100, log_eps = log(0.001), log_err = log(0.001)),
  lower_bounds = c(ngen = 1, log_eps = log(1e-12), log_err = log(1e-12)),
  upper_bounds = c(ngen = 1000, log_eps = log(0.2), log_err = log(0.2)),
  max_iter = 10,
  k = 20,
  npop = 600,
  ncores = 1,
  epsilon = 0.001,
  trace = TRUE
)
}
\arguments{
\item{tree_fit}{phylogenetic tree}

\item{amat}{alternative allele count matrix}

\item{dmat}{total depth matrix}

\item{initial_params}{initial parameter values (ngen, log_eps, log_err)}

\item{lower_bounds}{lower bounds for parameters in transformed space}

\item{upper_bounds}{upper bounds for parameters in transformed space}

\item{max_iter}{maximum number of iterations}

\item{k}{number of clusters for likelihood computation}

\item{npop}{population size for likelihood computation}

\item{ncores}{number of cores to use}

\item{epsilon}{convergence threshold}

\item{trace}{whether to return trace of parameter values}
}
\value{
parameter values or list of parameter values and trace
}
\description{
Fit tree parameters using EM algorithm
}
