% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitodrift.R
\name{fit_params_mcmc}
\alias{fit_params_mcmc}
\title{Fit tree parameters using MCMC}
\usage{
fit_params_mcmc(
  tree_fit,
  amat,
  dmat,
  initial_params = c(ngen = 100, log_eps = log(0.001), log_err = log(0.001)),
  lower_bounds = c(ngen = 1, log_eps = log(1e-18), log_err = log(1e-18)),
  upper_bounds = c(ngen = 1000, log_eps = log(0.2), log_err = log(0.2)),
  nsteps = 500,
  nchains = 1,
  outfile = NULL,
  npop = 600,
  k = 20,
  ncores = 1
)
}
\arguments{
\item{tree_fit}{phylogenetic tree}

\item{amat}{alternative allele count matrix}

\item{dmat}{total depth matrix}

\item{initial_params}{initial parameter values (ngen, log_eps, log_err)}

\item{lower_bounds}{lower bounds for parameters in transformed space}

\item{upper_bounds}{upper bounds for parameters in transformed space}

\item{nsteps}{number of MCMC steps}

\item{nchains}{number of MCMC chains}

\item{outfile}{file to save MCMC result}

\item{npop}{population size for likelihood computation}

\item{k}{number of clusters for likelihood computation}

\item{ncores}{number of cores to use}

\item{burnin}{number of MCMC steps to discard as burnin}
}
\value{
MCMC result object from fmcmc
}
\description{
Estimates tree parameters (number of generations, error rates) using MCMC sampling
with the fmcmc package.
}
