% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitodrift.R
\name{get_leaf_liks_mat_old}
\alias{get_leaf_liks_mat_old}
\title{get leaft likelihoods given mutation data in matrix format}
\usage{
get_leaf_liks_mat_old(amat, dmat, vafs, eps = 0, ncores = 1, log = FALSE)
}
\arguments{
\item{amat}{matrix of allele counts}

\item{dmat}{matrix of total counts}

\item{vafs}{vector of VAFs}

\item{eps}{error rate}

\item{log}{whether to return log likelihoods}
}
\value{
list of likelihood matrices
}
\description{
get leaft likelihoods given mutation data in matrix format
}
