% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_params.R
\name{get_param_lik_cpp}
\alias{get_param_lik_cpp}
\title{Compute the likelihood of a tree given model parameters}
\usage{
get_param_lik_cpp(tree_fit, amat, dmat, ngen, err, eps, npop = 600, k = 20)
}
\arguments{
\item{tree_fit}{phylogenetic tree}

\item{ngen}{number of generations}

\item{err}{error rate}

\item{eps}{mutation rate}

\item{npop}{population size}

\item{k}{number of clusters for likelihood computation}
}
\value{
log-likelihood
}
\description{
Compute the likelihood of a tree given model parameters
}
