% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{get_slide_freq_table}
\alias{get_slide_freq_table}
\title{Optimized version of \code{get_slide_freq_table} from rwty}
\usage{
get_slide_freq_table(
  tree.list,
  burnin = 0,
  window.size = 20,
  gens.per.tree = 1
)
}
\description{
Optimized version of \code{get_slide_freq_table} from rwty
}
