% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitodrift.R
\name{get_transition_mat_wf_hmm}
\alias{get_transition_mat_wf_hmm}
\title{Get the transition matrix for WF model with HMM}
\usage{
get_transition_mat_wf_hmm(k, eps, N, ngen)
}
\arguments{
\item{k}{number of VAF bins}

\item{eps}{error rate}

\item{N}{population size}

\item{ngen}{number of generations}
}
\value{
transition matrix
}
\description{
Get the transition matrix for WF model with HMM
}
