% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitodrift.R
\name{get_transition_mat_wf_hmm_wrapper}
\alias{get_transition_mat_wf_hmm_wrapper}
\title{Wrapper function to interpolate non-integer generations}
\usage{
get_transition_mat_wf_hmm_wrapper(k, eps, N, ngen)
}
\arguments{
\item{k}{number of VAF bins}

\item{eps}{error rate}

\item{N}{population size}

\item{ngen}{number of generations}
}
\value{
transition matrix
}
\description{
Wrapper function to interpolate non-integer generations
}
