% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logSumExp}
\alias{logSumExp}
\title{logSumExp function for a vector}
\usage{
logSumExp(x)
}
\arguments{
\item{x}{NumericVector}
}
\value{
double logSumExp of x
}
\description{
logSumExp function for a vector
}
\keyword{internal}
