% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_to_mat}
\alias{long_to_mat}
\title{Convert a long format mutation data frame to a matrix}
\usage{
long_to_mat(mut_dat_long, variable)
}
\arguments{
\item{mut_dat_long}{A data frame with columns: variant, cell, variable}

\item{variable}{The variable to convert to a matrix}
}
\value{
A matrix of values
}
\description{
Convert a long format mutation data frame to a matrix
}
