% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitodrift.R
\name{make_rooted_nj}
\alias{make_rooted_nj}
\title{Make a rooted NJ tree}
\usage{
make_rooted_nj(vmat, dist_method = "manhattan")
}
\arguments{
\item{vmat}{A matrix of cell-by-variable values}

\item{dist_method}{The distance method to use}
}
\value{
A phylo object
}
\description{
Make a rooted NJ tree
}
