% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_to_long}
\alias{mat_to_long}
\title{Convert allele count matrix (amat) and total count matrix (dmat) to long format}
\usage{
mat_to_long(amat, dmat)
}
\arguments{
\item{amat}{Allele count matrix with variants as rows and cells as columns}

\item{dmat}{Total count matrix with variants as rows and cells as columns}
}
\value{
A data.table in long format with columns: variant, cell, a (allele count), d (total count)
}
\description{
Convert allele count matrix (amat) and total count matrix (dmat) to long format
}
