% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phylo_to_gtree}
\alias{phylo_to_gtree}
\title{Convert a \code{phylo} object to a \code{tbl_graph}}
\usage{
phylo_to_gtree(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
A \code{tbl_graph} with \code{nodes} and \code{edges} from the input tree.
}
\description{
Builds a \code{tbl_graph} with tip nodes labeled by \code{phy$tip.label} and internal
nodes labeled from \code{phy$node.label}. If internal labels are missing, they are
created via \code{add_node_names()}.
}
