% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{smooth_features_phylo}
\alias{smooth_features_phylo}
\title{Smooth feature matrix over phylogenetic neighbors}
\usage{
smooth_features_phylo(feature_mat, phylo, k = 5, ties_method = "random")
}
\arguments{
\item{feature_mat}{A matrix with features as rows and cells as columns}

\item{phylo}{A phylo object}

\item{k}{Number of nearest neighbors (including self) to smooth over}

\item{ties_method}{Method to handle ties in distance: "min" (include all ties, default), "random", or "first"}
}
\value{
A smoothed matrix
}
\description{
Smooth feature matrix over phylogenetic neighbors
}
