% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitodrift.R
\name{get_transition_mat_wf_hmm}
\alias{get_transition_mat_wf_hmm}
\title{Get the transition matrix for WF model with HMM (with caching)
TODO: add log option for small probabilities}
\usage{
get_transition_mat_wf_hmm(k, eps, N, ngen, safe = FALSE)
}
\arguments{
\item{k}{number of VAF bins}

\item{eps}{Variant detection error rate}

\item{N}{population size}

\item{ngen}{number of generations}

\item{safe}{whether to add small probability to avoid 0s}
}
\value{
transition matrix
}
\description{
Get the transition matrix for WF model with HMM (with caching)
TODO: add log option for small probabilities
}
\keyword{internal}
